Deprecate IsolatedTest in favor of FluentStandalone and FluentStandaloneRunnable
Close #346
